MarqBridge-main/MarqBridge-main/MarqBridge-main/MarqBridge-main/MarqBridge-ddd3570bc219e36e720f0876100674069e916276/marqbridge/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
websockets==12.0
python-dotenv==1.0.1
httpx==0.27.0
ccxt==4.2.90
anthropic==0.25.0
pydantic==2.6.4
pydantic-settings==2.0.3
sqlalchemy==2.0.29
aiosqlite==0.20.0
alembic==1.13.1
pytz==2024.1
stripe==9.0.0
